BUGFIX: Extract Node Variant creation to Node Duplication Service and add handling for tags of duplicated node
