+++ Neos.Neos/Classes/Controller/Service/NodesController.php
@@ -16,19 +16,13 @@ namespace Neos\Neos\Controller\Service;
 
 use Neos\ContentRepository\Core\ContentRepository;
 use Neos\ContentRepository\Core\DimensionSpace\DimensionSpacePoint;
-use Neos\ContentRepository\Core\DimensionSpace\OriginDimensionSpacePoint;
-use Neos\ContentRepository\Core\Feature\NodeVariation\Command\CreateNodeVariant;
 use Neos\ContentRepository\Core\NodeType\NodeTypeNames;
 use Neos\ContentRepository\Core\Projection\ContentGraph\AbsoluteNodePath;
-use Neos\ContentRepository\Core\Projection\ContentGraph\ContentSubgraphInterface;
-use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\FindChildNodesFilter;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\FindDescendantNodesFilter;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\NodeType\ExpandedNodeTypeCriteria;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\NodeType\NodeTypeCriteria;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\SearchTerm\SearchTerm;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\SearchTerm\SearchTermMatcher;
-use Neos\ContentRepository\Core\Projection\ContentGraph\NodeAggregate;
-use Neos\ContentRepository\Core\Projection\ContentGraph\VisibilityConstraints;
 use Neos\ContentRepository\Core\SharedModel\Node\NodeAddress;
 use Neos\ContentRepository\Core\SharedModel\Node\NodeAggregateId;
 use Neos\ContentRepository\Core\SharedModel\Workspace\WorkspaceName;
@@ -38,6 +32,7 @@ use Neos\Flow\Mvc\Controller\ActionController;
 use Neos\Flow\Property\PropertyMapper;
 use Neos\FluidAdaptor\View\TemplateView;
 use Neos\Neos\Controller\BackendUserTranslationTrait;
+use Neos\Neos\Domain\Service\NodeDuplicationService;
 use Neos\Neos\Domain\Service\NodeTypeNameFactory;
 use Neos\Neos\FrontendRouting\SiteDetection\SiteDetectionResult;
 use Neos\Neos\Ui\Domain\Service\NodePropertyConverterService;
@@ -57,6 +52,12 @@ class NodesController extends ActionController
      * @var NodePropertyConverterService
      */
     protected $nodePropertyConverterService;
+
+    /**
+     * @Flow\Inject
+     * @var NodeDuplicationService
+     */
+    protected $nodeDuplicationService;
 
     /**
      * @Flow\Inject
@@ -286,7 +287,7 @@ class NodesController extends ActionController
         $targetSubgraph = $contentRepository->getContentSubgraph($workspaceName, $targetDimensionSpacePoint);
 
         if ($mode === 'adoptFromAnotherDimension' || $mode === 'adoptFromAnotherDimensionAndCopyContent') {
-            $this->adoptNodeAndParents(
+            $this->nodeDuplicationService->adoptNodeAndParents(
                 $workspaceName,
                 $nodeAggregateId,
                 $sourceSubgraph,
@@ -351,114 +352,6 @@ class NodesController extends ActionController
                     );
                 }
             }
-        }
-    }
-
-    /**
-     * Adopt (translate) the given node and parents that are not yet visible to the given context
-     *
-     * @param WorkspaceName $workspaceName
-     * @param NodeAggregateId $nodeAggregateId
-     * @param ContentSubgraphInterface $sourceSubgraph
-     * @param ContentSubgraphInterface $targetSubgraph
-     * @param DimensionSpacePoint $targetDimensionSpacePoint
-     * @param ContentRepository $contentRepository
-     * @param boolean $copyContent true if the content from the nodes that are translated should be copied
-     * @return void
-     */
-    protected function adoptNodeAndParents(
-        WorkspaceName $workspaceName,
-        NodeAggregateId $nodeAggregateId,
-        ContentSubgraphInterface $sourceSubgraph,
-        ContentSubgraphInterface $targetSubgraph,
-        DimensionSpacePoint $targetDimensionSpacePoint,
-        ContentRepository $contentRepository,
-        bool $copyContent
-    ): void {
-        $identifiersFromRootlineToTranslate = [];
-        while (
-            $nodeAggregateId
-            && $targetSubgraph->findNodeById($nodeAggregateId) === null
-        ) {
-            $identifiersFromRootlineToTranslate[] = $nodeAggregateId;
-            $nodeAggregateId = $sourceSubgraph->findParentNode($nodeAggregateId)
-                ?->aggregateId;
-        }
-        // $identifiersFromRootlineToTranslate is now bottom-to-top; so we need to reverse
-        // them to know what we need to create.
-        // TODO: TEST THAT AUTO CREATED CHILD NODES WORK (though this should not have influence)
-
-        foreach (array_reverse($identifiersFromRootlineToTranslate) as $identifier) {
-            assert($identifier instanceof NodeAggregateId);
-            // NOTE: for creating node variants, we need to find the ORIGIN DSP
-            // of the source node (in order to unambiguously identify it);
-            // so we need to load it from the source subgraph
-            $sourceNode = $sourceSubgraph->findNodeById($identifier);
-            if (!$sourceNode) {
-                throw new \RuntimeException('Source node for Node Aggregate ID ' . $identifier->value
-                    . ' not found. This should never happen.', 1660905374);
-            }
-            $contentRepository->handle(
-                CreateNodeVariant::create(
-                    $workspaceName,
-                    $identifier,
-                    $sourceNode->originDimensionSpacePoint,
-                    OriginDimensionSpacePoint::fromDimensionSpacePoint($targetDimensionSpacePoint),
-                )
-            );
-
-            if ($copyContent === true) {
-                $contentNodeConstraint = NodeTypeCriteria::fromFilterString('!' . NodeTypeNameFactory::NAME_DOCUMENT);
-                $this->createNodeVariantsForChildNodes(
-                    $workspaceName,
-                    $identifier,
-                    $contentNodeConstraint,
-                    $sourceSubgraph,
-                    $targetSubgraph,
-                    $targetDimensionSpacePoint,
-                    $contentRepository
-                );
-            }
-        }
-    }
-
-    private function createNodeVariantsForChildNodes(
-        WorkspaceName $workspaceName,
-        NodeAggregateId $parentNodeId,
-        NodeTypeCriteria $constraints,
-        ContentSubgraphInterface $sourceSubgraph,
-        ContentSubgraphInterface $targetSubgraph,
-        DimensionSpacePoint $targetDimensionSpacePoint,
-        ContentRepository $contentRepository
-    ): void {
-        foreach (
-            $sourceSubgraph->findChildNodes(
-                $parentNodeId,
-                FindChildNodesFilter::create(nodeTypes: $constraints)
-            ) as $childNode
-        ) {
-            if ($childNode->classification->isRegular()) {
-                // Tethered nodes' variants are automatically created when the parent is translated.
-                // TODO: DOES THIS MAKE SENSE?
-                $contentRepository->handle(
-                    CreateNodeVariant::create(
-                        $workspaceName,
-                        $childNode->aggregateId,
-                        $childNode->originDimensionSpacePoint,
-                        OriginDimensionSpacePoint::fromDimensionSpacePoint($targetDimensionSpacePoint),
-                    )
-                );
-            }
-
-            $this->createNodeVariantsForChildNodes(
-                $workspaceName,
-                $childNode->aggregateId,
-                $constraints,
-                $sourceSubgraph,
-                $targetSubgraph,
-                $targetDimensionSpacePoint,
-                $contentRepository
-            );
         }
     }
 }
